requirements.txt
asgiref==3.3.1
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-admin-ordering==0.13.1
django-adminlte3==0.1.6
django-import-export==2.4.0
django-jazzmin==2.4.4
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-registration==3.1.1
django-registration-redux==2.8
django-rest-framework==0.1.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
#numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
#pandas==1.1.5
#pysqlite3-binary==0.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
#uWSGI==2.0.19.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
